7 write unit test (#22)
* fix: cassert is not included

* add test with GoogleTest

* update: reduce ci time

* remove using pollution

* update: split declaration and definition

* add: test for RefValues

* fix generator cannot compile

* update: parallelize build
