+++ src/particle.hpp
@@ -66,30 +66,12 @@ public:
 	 * @param pos  position of the particle
 	 * @param vel 	velocity of the particle
 	 */
-	Particle(int id, ParticleType type, Eigen::Vector3d pos, Eigen::Vector3d vel) {
-		this->id       = id;
-		this->type     = type;
-		this->position = pos;
-		this->velocity = vel;
-	}
+	Particle(int id, ParticleType type, Eigen::Vector3d pos, Eigen::Vector3d vel);
 
 	/**
 	 * @brief calculate inverse of density
 	 * @param density density of the particle
 	 * @return inverse of density
 	 */
-	double inverseDensity(double& density) const {
-		switch (type) {
-		case ParticleType::Ghost:
-			return std::numeric_limits<double>::infinity();
-
-		case ParticleType::Fluid:
-			return 1 / density;
-
-		case ParticleType::Wall:
-		case ParticleType::DummyWall:
-		default:
-			return 0;
-		}
-	}
+	double inverseDensity(double& density) const;
 };
